@@ -1,5 +1,23 @@
-module.exports = {
-  docs: [
+/**
+ * Creating a sidebar enables you to:
+ - create an ordered group of docs
+ - render a sidebar for each doc of that group
+ - provide next/previous navigation
+
+ The sidebars can be generated from the filesystem, or explicitly defined here.
+
+ Create as many sidebars as you want.
+ */
+
+// @ts-check
+
+/** @type {import('@docusaurus/plugin-content-docs').SidebarsConfig} */
+const sidebars = {
+  // By default, Docusaurus generates a sidebar from the docs folder structure
+  // tutorialSidebar: [{type: 'autogenerated', dirName: '.'}],
+
+  // But you can create a sidebar manually
+  tutorialSidebar: [
     {'Smart Data Lake' : [
       'intro',
       'features',
@@ -34,40 +52,42 @@ module.exports = {
 
     ]},
     {'Reference' : [
-        'reference/build',
-        'reference/commandLine',
-        {'SDLB objects' : [
-          'reference/dataObjects',
-          'reference/actions',
-        ]},
-        {'Hocon Configuration' : [
-          'reference/hoconOverview',
-          'reference/hoconVariables',
-          'reference/hoconSecrets',
-        ]},
-        'reference/dag',
-        'reference/schema',
-        'reference/schema-viewer-navigation',
-        'reference/dataQuality',
-        'reference/executionPhases',
-        'reference/executionEngines',
-        'reference/executionModes',
-        'reference/transformations',
-        //'reference/schemaEvolution',
-        //'reference/housekeeping',
-        'reference/streaming',
-        {'Deployment' : [
-          'reference/deploymentOptions',
-          'reference/deploy-microsoft-azure',
-        ]},
-        'reference/testing',
-        'reference/troubleshooting',
-        //'reference/glossary'
-    ]},
-    {
-      type: 'link',
-      label: 'Configuration Schema Viewer', // The link label
-      href: '/json-schema-viewer', // The internal path
-    },
-  ],
+      'reference/build',
+      'reference/commandLine',
+      {'SDLB objects' : [
+        'reference/dataObjects',
+        'reference/actions',
+      ]},
+      {'Hocon Configuration' : [
+        'reference/hoconOverview',
+        'reference/hoconVariables',
+        'reference/hoconSecrets',
+      ]},
+      'reference/dag',
+      'reference/schema',
+      'reference/schema-viewer-navigation',
+      'reference/dataQuality',
+      'reference/executionPhases',
+      'reference/executionEngines',
+      'reference/executionModes',
+      'reference/transformations',
+      //'reference/schemaEvolution',
+      //'reference/housekeeping',
+      'reference/streaming',
+      {'Deployment' : [
+        'reference/deploymentOptions',
+        'reference/deploy-microsoft-azure',
+      ]},
+      'reference/testing',
+      'reference/troubleshooting',
+      //'reference/glossary'
+  ]},
+  {
+    type: 'link',
+    label: 'Configuration Schema Viewer', // The link label
+    href: '/json-schema-viewer', // The internal path
+  },
+  ]
 };
+
+export default sidebars;
